added responsive and organized according to the general structure
