+++ package.json
@@ -6,7 +6,9 @@
     "@reduxjs/toolkit": "^1.5.1",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "email-validator": "^2.0.4",
     "formik": "^2.2.9",
+    "formik-field": "^1.6.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.3",
