+++ src/App.jsx
@@ -1,9 +1,16 @@
-import Account from './account';
+import { BrowserRouter } from 'react-router-dom';
+
+import './App.css';
+import Login from './components/Login/Login';
+import Signup from './components/Signup/Signup';
 
 function App() {
   return (
     <div className="App">
-      <Account />
+       <BrowserRouter>
+      <Login />
+      <Signup />
+      </BrowserRouter>
     </div>
   );
 }
