+++ src/components/Signup/tests/Signup.test.js
@@ -1,15 +1,14 @@
-import { BrowserRouter } from 'react-router-dom';
 import renderer from 'react-test-renderer';
-import Signup from '../components/Signup';
-
+import { BrowserRouter as Router } from 'react-router-dom';
+import Signup from '../Signup';
 
 describe('Signup', () => {
   it('matches', () => {
     const tree = renderer
       .create(
-        <BrowserRouter>
+        <Router>
           <Signup />
-        </BrowserRouter>,
+        </Router>,
       )
       .toJSON();
     expect(tree).toMatchSnapshot();
