+++ src/routers.js
@@ -1,0 +1,2 @@
+export const Login = '/Login';
+export const Signup = '/Signup';
